Started work on a model builder
