+++ src/render/model/handle.rs
@@ -1,63 +1,7 @@
-use super::Model;
-use cgmath::{Euler, Matrix4, Rad, SquareMatrix, Vector3, Zero};
+use super::{Model, ModelData};
+use cgmath::{Euler, Rad, Vector3};
 use parking_lot::RwLock;
-use std::sync::{
-    atomic::{AtomicU64, Ordering},
-    mpsc::Sender,
-    Arc,
-};
-
-/// Data of a model. This is behind an `Arc<RwLock<>>` so that the engine can keep a copy and check the latest values.
-///
-/// For an example on how to use this, see the example in the root of this module. This is the value passed in `ModelHandle::modify`.
-pub struct ModelData {
-    pub(crate) id: u64,
-    pub(crate) model: Arc<Model>,
-    /// The current position in the world that this model exists at.
-    pub position: Vector3<f32>,
-    /// The rotation of this model, in euler angles.
-    pub rotation: Euler<Rad<f32>>,
-    /// The scale of this model.
-    pub scale: f32,
-
-    pub group_matrices: Vec<Matrix4<f32>>,
-}
-
-impl std::fmt::Debug for ModelData {
-    fn fmt(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
-        fmt.debug_struct("ModelData")
-            .field("position", &self.position)
-            .field("rotation", &self.rotation)
-            .field("scale", &self.scale)
-            .finish()
-    }
-}
-
-impl ModelData {
-    pub(self) fn new(model: Arc<Model>) -> (u64, Arc<RwLock<Self>>) {
-        static ID: AtomicU64 = AtomicU64::new(0);
-        let id = ID.fetch_add(1, Ordering::Relaxed);
-        let index_group_count = model.indices.len();
-        (
-            id,
-            Arc::new(RwLock::new(Self {
-                id,
-                model,
-                position: Vector3::zero(),
-                rotation: Euler::new(Rad(0.0), Rad(0.0), Rad(0.0)),
-                scale: 1.0,
-                group_matrices: (0..index_group_count)
-                    .map(|_| Matrix4::identity())
-                    .collect(),
-            })),
-        )
-    }
-    pub(crate) fn matrix(&self) -> Matrix4<f32> {
-        Matrix4::from_translation(self.position)
-            * Matrix4::from(self.rotation)
-            * Matrix4::from_scale(self.scale)
-    }
-}
+use std::sync::{mpsc::Sender, Arc};
 
 /// A handle to the model that was loaded. This can be used to move the model around in the world.
 ///
@@ -168,7 +112,7 @@ impl Drop for ModelHandle {
     }
 }
 
-pub(crate) enum ModelHandleMessage {
+pub enum ModelHandleMessage {
     NewClone(u64, Arc<RwLock<ModelData>>),
     Dropped(u64),
 }
